[Clang][CodeGen][NFC] Refactor CGCXXABI classes for code sharing
This refactors the CGCXXABI and ItaniumCXXABI classes in LLVM IR codegen
and their CIR codegen counterparts so that common code can be shared
between the two.

This is a small first step towards better code sharing between LLVM IR
codegen and CIR codegen. We should be able to expand this significantly
as CIR codegen matures.

Note that some other code can very nearly, but not quite be shared at this
point. For example `AddedStructorArgs` and its uses are nearly identical
between LLVM IR and CIR codegen, but the struct contains an `llvm::Value*`
member in LLVM IR codegen where it conatins an `mlir::Value` member in
CIR codegen. We will need additional refactoring to handle cases like that.
